split frontend
